// model/createUserResponse.ts
/**
 * ManageSpace Auth API
 * ManageSpace Auth API Documentation
 *
 * The version of the OpenAPI document: 0.0.11-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CodeDeliveryDetailsResponse } from './codeDeliveryDetailsResponse';

export class CreateUserResponse {
    /**
    * User Confirmation Status
    */
    'userConfirmed': boolean;
    'codeDeliveryDetails'?: CodeDeliveryDetailsResponse;
    /**
    * User ID
    */
    'userSub': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "userConfirmed",
            "baseName": "UserConfirmed",
            "type": "boolean"
        },
        {
            "name": "codeDeliveryDetails",
            "baseName": "CodeDeliveryDetails",
            "type": "CodeDeliveryDetailsResponse"
        },
        {
            "name": "userSub",
            "baseName": "UserSub",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateUserResponse.attributeTypeMap;
    }
}